refactor : Point Interface화
