+++ src/main/java/nextstep/ladder/domain/lines/point/Point.java
@@ -1,0 +1,12 @@
+package nextstep.ladder.domain.lines.point;
+
+import nextstep.ladder.domain.lines.point.impl.PointImpl;
+
+public interface Point {
+
+    PointImpl decideNextPoint(Boolean randomBoolean);
+
+    boolean isPointFalse();
+
+    boolean isPointTrue();
+}
